update cleanup script

+++ benchmarks/clean-up-projects.sh
@@ -6,18 +6,24 @@ if ! [ -n "$1" ]; then
     exit 1
 fi
 
-cd $1
+if [ -d $1 ]; then
+    echo "Deleting old logs in $1"
 
-find . -name "inferTiming\.log" | xargs rm -f
-find . -name "infer\.log" | xargs rm -f
-find . -name "*\.jaif" | xargs rm -f
-find . -name "*\.class" | xargs rm -f
-find . -name "solutions\.txt" | xargs rm -f
-find . -name "statistics\.txt" | xargs rm -f
-find . -name "unsatConstraints\.txt" | xargs rm -f
-find . -name "slots\.smt" | xargs rm -f
-find . -name "constraints\.smt" | xargs rm -f
-find . -name "z3Constraints\.smt" | xargs rm -f
-find . -name "z3ConstraintsUnsatCore\.smt" | xargs rm -f
-find . -name "z3ConstraintsGlob\.smt" | xargs rm -f
-find . -name "gjeConstraints*\.gje" | xargs rm -f
+    cd $1
+
+    find . -name "inferTiming\.log" | xargs rm -f
+    find . -name "infer\.log" | xargs rm -f
+    find . -name "*\.jaif" | xargs rm -f
+    find . -name "*\.class" | xargs rm -f
+    find . -name "solutions\.txt" | xargs rm -f
+    find . -name "statistics\.txt" | xargs rm -f
+    find . -name "unsatConstraints\.txt" | xargs rm -f
+    find . -name "slots\.smt" | xargs rm -f
+    find . -name "constraints\.smt" | xargs rm -f
+    find . -name "z3Constraints\.smt" | xargs rm -f
+    find . -name "z3ConstraintsUnsatCore\.smt" | xargs rm -f
+    find . -name "z3ConstraintsGlob\.smt" | xargs rm -f
+    find . -name "gjeConstraints*\.gje" | xargs rm -f
+
+    echo "Done"
+fi
